--- src/main/java/com/example/bookstores/Model/Book.java
@@ -1,4 +1,0 @@
-package com.example.bookstores.Model;
-
-public class Book {
-}
