--- src/main/java/com/example/bookstores/Repository/BookRepository.java
@@ -1,4 +1,0 @@
-package com.example.bookstores.Repository;
-
-public class BookRepository {
-}
