--- src/main/java/com/example/bookstores/Repository/IBookRepository.java
@@ -1,4 +1,0 @@
-package com.example.bookstores.Repository;
-
-public interface IBookRepository {
-}
