--- src/main/java/com/example/bookstores/Service/BookService.java
@@ -1,4 +1,0 @@
-package com.example.bookstores.Service;
-
-public class BookService {
-}
