--- src/main/java/com/example/bookstores/Service/IBookService.java
@@ -1,4 +1,0 @@
-package com.example.bookstores.Service;
-
-public interface IBookService {
-}
