+++ src/main/java/com/example/bookstores/model/Book.java
@@ -1,0 +1,4 @@
+package com.example.bookstores.model;
+
+public class Book {
+}
