+++ src/main/java/com/example/bookstores/repository/BookRepository.java
@@ -1,0 +1,4 @@
+package com.example.bookstores.repository;
+
+public class BookRepository {
+}
