+++ src/main/java/com/example/bookstores/repository/CustomerRepository.java
@@ -1,4 +1,4 @@
-package com.example.bookstores.Repository;
+package com.example.bookstores.repository;
 
 public class CustomerRepository implements ICustomerRepository{
