+++ src/main/java/com/example/bookstores/repository/IBookRepository.java
@@ -1,0 +1,4 @@
+package com.example.bookstores.repository;
+
+public interface IBookRepository {
+}
