+++ src/main/java/com/example/bookstores/repository/ICustomerRepository.java
@@ -1,0 +1,4 @@
+package com.example.bookstores.repository;
+
+public interface ICustomerRepository {
+}
