+++ src/main/java/com/example/bookstores/service/BookService.java
@@ -1,0 +1,4 @@
+package com.example.bookstores.service;
+
+public class BookService {
+}
