+++ src/main/java/com/example/bookstores/service/IBookService.java
@@ -1,0 +1,4 @@
+package com.example.bookstores.service;
+
+public interface IBookService {
+}
